Support setting a default new-shell config
For users that only use a single agent, they might not want to always have to be prompted to select
from one of a list of mostly-irrelevant-to-them agent configurations.

For these users, we provide a `agent-shell-new-shell-config` with which they can define a specific
agent configuration to unconditionally use for all new agent-shell sessions.

For example, the following will allow users to unconditionally use Claude Code by default:

```elisp
(setq agent-shell-new-shell-config (agent-shell-anthropic-make-claude-code-config))
```